# service/docs/swagger.yaml
basePath: /
definitions:
  controller.CreateMessageRequest:
    properties:
      text:
        type: string
    type: object
  controller.CreateMessageResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      room_id:
        type: integer
      text:
        type: string
      username:
        type: string
    type: object
  controller.CreateRoomRequest:
    properties:
      name:
        type: string
    type: object
  controller.CreateRoomResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  controller.GetRoomResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  controller.ListMessageResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/controller.MessageView'
        type: array
    type: object
  controller.ListRoomResponse:
    properties:
      rooms:
        items:
          $ref: '#/definitions/controller.RoomView'
        type: array
    type: object
  controller.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controller.LoginResponse:
    properties:
      code:
        type: integer
      expire:
        type: string
      token:
        type: string
    type: object
  controller.MessageView:
    properties:
      created_at:
        type: string
      id:
        type: integer
      room_id:
        type: integer
      text:
        type: string
      username:
        type: string
    type: object
  controller.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  controller.RegisterResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      username:
        type: string
    required:
    - email
    - username
    type: object
  controller.RoomView:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:8001
info:
  contact:
    email: hernanrocha93(at)gmail.com
    name: Hernan Rocha
  description: This is a simple bot-based chat.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger FinChat API
  version: "1.0"
paths:
  /api/v1/rooms:
    get:
      description: List Rooms in database
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ListRoomResponse'
            type: object
      summary: List Rooms
      tags:
      - Rooms
    post:
      description: Create Room in database
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Room Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controller.CreateRoomRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CreateRoomResponse'
            type: object
      summary: Create Room
      tags:
      - Rooms
  /api/v1/rooms/{id}:
    get:
      description: Get Room by ID
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetRoomResponse'
            type: object
      summary: Get Room
      tags:
      - Rooms
  /api/v1/rooms/{id}/messages:
    get:
      description: List last Room Messages in database
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ListMessageResponse'
            type: object
      summary: List Room Messages
      tags:
      - Messages
    post:
      description: Create Message in database
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controller.CreateMessageRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CreateMessageResponse'
            type: object
      summary: Create Message
      tags:
      - Messages
  /login:
    post:
      description: Login with Username and Password
      parameters:
      - description: Login Credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/controller.LoginRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.LoginResponse'
            type: object
      summary: Login
      tags:
      - Authentication
  /register:
    post:
      description: Register User in database
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.RegisterResponse'
            type: object
      summary: Register User
      tags:
      - Authentication
swagger: "2.0"